backend/recommend-service/requirements.txt
anyio==3.6.2
autopep8==2.0.2
click==8.1.3
cmake==3.26.1
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.95.0
filelock==3.11.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pandas==2.0.0
psycopg2==2.9.6
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
